pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image-gen"
version = "0.1.0"
description = "Generate or edit images using OpenAI GPT or Google Gemini APIs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Wolf" }]
keywords = ["image-generation", "openai", "gemini", "cli", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Typing :: Typed",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
gemini = ["google-genai>=0.1.0", "Pillow>=10.0.0"]
all = ["image-gen[openai,gemini]"]
dev = [
    "image-gen[all]",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
image-gen = "image_gen.cli:main"

[project.urls]
Homepage = "https://github.com/wolf/image-gen"
Repository = "https://github.com/wolf/image-gen"
Issues = "https://github.com/wolf/image-gen/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/image_gen"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["image_gen"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
files = ["src/image_gen"]

[[tool.mypy.overrides]]
module = [
    "openai.*",
    "google.*",
    "PIL.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "image_gen.backends.openai_backend",
    "image_gen.backends.gemini_backend",
]
# These modules interact heavily with external SDK APIs that have incomplete/dynamic type stubs
disallow_untyped_defs = false
disable_error_code = ["call-overload", "attr-defined", "valid-type", "arg-type"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
]

[tool.coverage.run]
source = ["src/image_gen"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
